Update git clone command in README to point to main nomic repo=
I'm not sure if it was intentional that the build instructions tell the user to clone `manyoso/gpt4all-chat.git`?

But I would think this should be cloning the main repo at `nomic-ai/gpt4all-chat` instead.  Otherwise users following this command might get changes not yet merged into the main repo, which could be confusing.